DEV: move gen_random_xarray to testing module and use it in another test
