+++ pylabianca/testing.py
@@ -1,3 +1,5 @@
+from string import ascii_lowercase
+
 import numpy as np
 import pytest
 
@@ -24,3 +26,31 @@ def spk_epochs(ft_data):
                            keep_timestamps=True)
     )
     return spk_epo_test
+
+
+def gen_random_xarr(n_cells, n_trials, n_times, per_cell_coord=False):
+    import xarray as xr
+
+    letters = np.array(list(ascii_lowercase))
+    dim_names = ['cell', 'trial', 'time']
+    time = np.linspace(-0.5, 1.5, num=n_times)
+    data = np.random.rand(n_cells, n_trials, n_times)
+    cell_names = [''.join(np.random.choice(letters, 5))
+                  for _ in range(n_cells)]
+
+    xarr = xr.DataArray(
+        data, dims=dim_names,
+        coords={'cell': cell_names,
+                'trial': np.arange(n_trials),
+                'time': time}
+    )
+
+    if per_cell_coord:
+        prefs = np.zeros((n_cells, n_trials), dtype=int)
+        for cell_idx in range(n_cells):
+            this_prefs = np.random.choice([0, 1, 2], size=n_trials)
+            prefs[cell_idx, :] = this_prefs
+
+        xarr = xarr.assign_coords(preferred=(('cell', 'trial'), prefs))
+
+    return xarr
